Port new page-streaming code to REST APIs.
