+++ src/Google.Api.Gax.Rest/PageStreamingAsync.cs
@@ -5,7 +5,7 @@
  * https://developers.google.com/open-source/licenses/bsd
  */
 
-using Google.Protobuf;
+using Google.Apis.Requests;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -15,7 +15,7 @@ using System.Threading.Tasks;
 // Interfaces and implementations of asynchronous page streaming. These are gathered into a single file
 // for convenience.
 
-namespace Google.Api.Gax
+namespace Google.Api.Gax.Rest
 {
     /// <summary>
     /// An asynchronous sequence of resources obtained via API responses. Application code
@@ -89,29 +89,30 @@ namespace Google.Api.Gax
     /// <typeparam name="TResponse">The API response type. Each response contains a page of resources.</typeparam>
     /// <typeparam name="TResource">The resource type contained within the response.</typeparam>
     public sealed class PagedAsyncEnumerable<TRequest, TResponse, TResource> : IPagedAsyncEnumerable<TResponse, TResource>
-        where TRequest : class, IPageRequest, IMessage<TRequest>
-        where TResponse : class, IPageResponse<TResource>, IMessage<TResponse>
+        where TRequest : class, IClientServiceRequest<TResponse>
+        where TResponse : class
     {
         private readonly IResponseAsyncEnumerable<TResponse, TResource> _pages;
+        private readonly IPageManager<TRequest, TResponse, TResource> _pageManager;
 
         /// <summary>
-        /// Creates a new lazily-evaluated asynchronous sequence from the given API call, initial request, and call settings.
+        /// Creates a new lazily-evaluated sequence from the given API call, initial request, and call settings.
         /// </summary>
-        /// <remarks>The request is cloned each time the sequence is evaluated.</remarks>
-        /// <param name="apiCall">The API call made each time a page is required.</param>
-        /// <param name="request">The initial request.</param>
-        /// <param name="callSettings">The settings to apply to each API call.</param>
-        public PagedAsyncEnumerable(ApiCall<TRequest, TResponse> apiCall,
-            TRequest request, CallSettings callSettings)
+        /// <param name="requestProvider">A factory used to create an initial request each time the sequence is iterated over.</param>
+        /// <param name="pageManager">A manager to work with the requests and responses.</param>
+        public PagedAsyncEnumerable(Func<TRequest> requestProvider,
+            IPageManager<TRequest, TResponse, TResource> pageManager)
         {
-            _pages = new ResponseAsyncEnumerable<TRequest, TResponse, TResource>(apiCall, request, callSettings);
+            _pages = new ResponseAsyncEnumerable<TRequest, TResponse, TResource>(requestProvider, pageManager);
+            _pageManager = pageManager;
         }
 
         /// <inheritdoc/>
         public IResponseAsyncEnumerable<TResponse, TResource> AsPages() => _pages;
 
         /// <inheritdoc/>
-        public IAsyncEnumerator<TResource> GetEnumerator() => _pages.SelectMany(page => page.ToAsyncEnumerable()).GetEnumerator();
+        public IAsyncEnumerator<TResource> GetEnumerator() =>
+            _pages.SelectMany(page => _pageManager.GetResourcesEmptyIfNull(page).ToAsyncEnumerable()).GetEnumerator();
     }
 
     /// <summary>
@@ -121,43 +122,39 @@ namespace Google.Api.Gax
     /// <typeparam name="TResponse">The API response type.</typeparam>
     /// <typeparam name="TResource">The resource type contained within the response.</typeparam>
     internal sealed class ResponseAsyncEnumerable<TRequest, TResponse, TResource> : IResponseAsyncEnumerable<TResponse, TResource>
-        where TRequest : class, IPageRequest, IMessage<TRequest>
-        where TResponse : class, IPageResponse<TResource>, IMessage<TResponse>
+        where TRequest : class, IClientServiceRequest<TResponse>
+        where TResponse : class
     {
-        private readonly CallSettings _callSettings;
-        private readonly TRequest _request;
-        private readonly ApiCall<TRequest, TResponse> _apiCall;
+        private readonly Func<TRequest> _requestProvider;
+        private readonly IPageManager<TRequest, TResponse, TResource> _pageManager;
 
-        public ResponseAsyncEnumerable(ApiCall<TRequest, TResponse> apiCall,
-            TRequest request, CallSettings callSettings)
+        public ResponseAsyncEnumerable(Func<TRequest> requestProvider,
+            IPageManager<TRequest, TResponse, TResource> pageManager)
         {
-            _callSettings = callSettings;
-            _request = request;
-            _apiCall = apiCall;
+            _requestProvider = GaxRestPreconditions.CheckNotNull(requestProvider, nameof(requestProvider));
+            _pageManager = GaxRestPreconditions.CheckNotNull(pageManager, nameof(pageManager));
         }
 
+
         public IResponseAsyncEnumerator<TResponse> GetEnumerator() =>
-            new ResponseAsyncEnumerator(_callSettings, _request.Clone(), _apiCall);
+            new ResponseAsyncEnumerator(_requestProvider(), _pageManager);
 
         IAsyncEnumerator<TResponse> IAsyncEnumerable<TResponse>.GetEnumerator() => GetEnumerator();
 
         /// <inheritdoc />
         public IAsyncEnumerable<FixedSizePage<TResource>> WithFixedSize(int pageSize) =>
-            new FixedPageSizeAsyncEnumerable(this, pageSize);
+            new FixedPageSizeAsyncEnumerable(this, _pageManager, pageSize);
 
         private class ResponseAsyncEnumerator : IResponseAsyncEnumerator<TResponse>
         {
-            private readonly CallSettings _callSettings;
-            private readonly ApiCall<TRequest, TResponse> _apiCall;
             private readonly TRequest _request; // This is mutated during iteration
+            private readonly IPageManager<TRequest, TResponse, TResource> _pageManager;
             private bool _finished;
 
-            public ResponseAsyncEnumerator(CallSettings callSettings,
-                TRequest request, ApiCall<TRequest, TResponse> apiCall)
+            public ResponseAsyncEnumerator(TRequest request, IPageManager<TRequest, TResponse, TResource> pageManager)
             {
-                _callSettings = callSettings;
                 _request = request;
-                _apiCall = apiCall;
+                _pageManager = pageManager;
             }
 
             public TResponse Current { get; private set; }
@@ -169,26 +166,21 @@ namespace Google.Api.Gax
                 {
                     return false;
                 }
-                CallSettings effectiveCallSettings = _callSettings;
-                if (cancellationToken != default(CancellationToken))
-                {
-                    effectiveCallSettings = new CallSettings(_callSettings) { CancellationToken = cancellationToken };
-                }
-                Current = await _apiCall.Async(_request, effectiveCallSettings);
-                var nextPageToken = Current.NextPageToken;
-                if (nextPageToken == "")
+                Current = await _request.ExecuteAsync(cancellationToken);
+                var nextPageToken = _pageManager.GetNextPageToken(Current);
+                if (nextPageToken == null)
                 {
                     _finished = true;
                 }
                 // Prepare the next request...
-                _request.PageToken = nextPageToken;
+                _pageManager.SetPageToken(_request, nextPageToken);
                 return true;
             }
 
             public Task<bool> MoveNext(int pageSize, CancellationToken cancellationToken)
             {
-                GaxPreconditions.CheckArgument(pageSize > 0, nameof(pageSize), "Must be greater than 0");
-                _request.PageSize = pageSize;
+                GaxRestPreconditions.CheckArgument(pageSize > 0, nameof(pageSize), "Must be greater than 0");
+                _pageManager.SetPageSize(_request, pageSize);
                 return MoveNext(cancellationToken);
             }
 
@@ -198,28 +190,36 @@ namespace Google.Api.Gax
         private class FixedPageSizeAsyncEnumerable : IAsyncEnumerable<FixedSizePage<TResource>>
         {
             private readonly IResponseAsyncEnumerable<TResponse, TResource> _source;
+            private readonly IPageManager<TRequest, TResponse, TResource> _pageManager;
             private readonly int _pageSize;
 
             internal FixedPageSizeAsyncEnumerable(
-                IResponseAsyncEnumerable<TResponse, TResource> source, int pageSize)
+                IResponseAsyncEnumerable<TResponse, TResource> source,
+                IPageManager<TRequest, TResponse, TResource> pageManager,
+                int pageSize)
             {
-                GaxPreconditions.CheckArgument(pageSize > 0, nameof(pageSize), "Must be greater than 0");
+                GaxRestPreconditions.CheckArgument(pageSize > 0, nameof(pageSize), "Must be greater than 0");
                 _source = source;
+                _pageManager = pageManager;
                 _pageSize = pageSize;
             }
 
             public IAsyncEnumerator<FixedSizePage<TResource>> GetEnumerator() =>
-                new Enumerator(_source.GetEnumerator(), _pageSize);
+                new Enumerator(_source.GetEnumerator(), _pageManager, _pageSize);
 
             private class Enumerator : IAsyncEnumerator<FixedSizePage<TResource>>
             {
                 private readonly IResponseAsyncEnumerator<TResponse> _enumerator;
+                private readonly IPageManager<TRequest, TResponse, TResource> _pageManager;
                 private readonly int _pageSize;
 
                 internal Enumerator(
-                    IResponseAsyncEnumerator<TResponse> enumerator, int pageSize)
+                    IResponseAsyncEnumerator<TResponse> enumerator,
+                    IPageManager<TRequest, TResponse, TResource> pageManager,
+                    int pageSize)
                 {
                     _enumerator = enumerator;
+                    _pageManager = pageManager;
                     _pageSize = pageSize;
                 }
 
@@ -237,17 +237,18 @@ namespace Google.Api.Gax
                         {
                             break;
                         }
-                        items.AddRange(_enumerator.Current);
+                        var resources = _pageManager.GetResourcesEmptyIfNull(_enumerator.Current);
+                        items.AddRange(resources);
                         if (items.Count > _pageSize)
                         {
                             // TODO: Better exception type?
                             throw new NotSupportedException("Invalid server response: " +
-                                $"requested {requestCount} items, received {_enumerator.Current.Count()} items");
+                                $"requested {requestCount} items, received {resources.Count()} items");
                         }
                     }
                     if (items.Count != 0)
                     {
-                        Current = new FixedSizePage<TResource>(items, _enumerator.Current.NextPageToken);
+                        Current = new FixedSizePage<TResource>(items, _pageManager.GetNextPageToken(_enumerator.Current));
                         return true;
                     }
                     Current = null;
