+++ test/Google.Api.Gax.Rest.Tests/PageStreamingTest.cs
@@ -161,7 +161,7 @@ namespace Google.Api.Gax.Rest
             var server = new FakeServer(pagedResource);
             var request = new PageStreamingRequest(server) { PageSize = 0 };
             var paged = server.PagedSync(request);
-            Assert.Equal(pagedResource.Resource, paged.Select(x => x.Items.ToArray()).ToArray());
+            Assert.Equal(pagedResource.Resource, paged.AsPages().Select(x => x.Items.ToArray()).ToArray());
         }
 
         [Theory, MemberData(nameof(s_naturalPages))]
@@ -170,7 +170,25 @@ namespace Google.Api.Gax.Rest
             var server = new FakeServer(pagedResource);
             var request = new PageStreamingRequest(server) { PageSize = 0 };
             var paged = server.PagedAsync(request);
-            Assert.Equal(pagedResource.Resource, await paged.Select(x => x.Items.ToArray()).ToArray());
+            Assert.Equal(pagedResource.Resource, await paged.AsPages().Select(x => x.Items.ToArray()).ToArray());
+        }
+
+        [Fact]
+        public void SpecifyPageTokenAtStart()
+        {
+            var server = new FakeServer(s_resourceA, 1);
+            var request = new PageStreamingRequest(server) { PageSize = 0, PageToken = "1:0" };
+            var paged = server.PagedSync(request);
+            Assert.Equal(s_resourceA.Resource.Skip(1), paged.AsPages().Select(x => x.Items.ToArray()));
+        }
+
+        [Fact]
+        public async Task SpecifyPageTokenAtStartAsync()
+        {
+            var server = new FakeServer(s_resourceA, 1);
+            var request = new PageStreamingRequest(server) { PageSize = 0, PageToken = "1:0" };
+            var paged = server.PagedAsync(request);
+            Assert.Equal(s_resourceA.Resource.Skip(1), await paged.AsPages().Select(x => x.Items.ToArray()).ToArray());
         }
 
         public static MatrixTheoryData<PagedResource, int> s_flatten = MatrixTheoryData.Create(
@@ -183,7 +201,7 @@ namespace Google.Api.Gax.Rest
             var server = new FakeServer(pagedResource);
             var request = new PageStreamingRequest(server) { PageSize = pageSize };
             var paged = server.PagedSync(request);
-            Assert.Equal(pagedResource.All, paged.Flatten());
+            Assert.Equal(pagedResource.All, paged);
         }
 
         [Theory, MemberData(nameof(s_flatten))]
@@ -192,7 +210,7 @@ namespace Google.Api.Gax.Rest
             var server = new FakeServer(pagedResource);
             var request = new PageStreamingRequest(server) { PageSize = pageSize };
             var paged = server.PagedAsync(request);
-            Assert.Equal(pagedResource.All, await paged.Flatten().ToArray());
+            Assert.Equal(pagedResource.All, await paged.ToArray());
         }
 
         public static MatrixTheoryData<PagedResource, int> s_fixedPageSize = MatrixTheoryData.Create(
@@ -205,7 +223,7 @@ namespace Google.Api.Gax.Rest
             var server = new FakeServer(pagedResource);
             var request = new PageStreamingRequest(server) { PageSize = pageSize };
             var paged = server.PagedSync(request);
-            var fixedSizePages = paged.WithFixedPageSize(pageSize).Select(page => page.ToArray()).ToArray();
+            var fixedSizePages = paged.AsPages().WithFixedSize(pageSize).Select(page => page.ToArray()).ToArray();
             var expectedPages = pagedResource.Fixed(pageSize);
             Assert.Equal(expectedPages, fixedSizePages);
         }
@@ -216,7 +234,7 @@ namespace Google.Api.Gax.Rest
             var server = new FakeServer(pagedResource);
             var request = new PageStreamingRequest(server) { PageSize = pageSize };
             var paged = server.PagedAsync(request);
-            var fixedSizePages = await paged.WithFixedPageSize(pageSize).Select(page => page.ToArray()).ToArray();
+            var fixedSizePages = await paged.AsPages().WithFixedSize(pageSize).Select(page => page.ToArray()).ToArray();
             var expectedPages = pagedResource.Fixed(pageSize);
             Assert.Equal(expectedPages, fixedSizePages);
         }
@@ -229,13 +247,13 @@ namespace Google.Api.Gax.Rest
             var request = new PageStreamingRequest(server) { PageSize = 0 };
             var paged = server.PagedSync(request);
             // Natural pages
-            Assert.Equal(1, paged.Count());
-            var page1 = paged.First();
+            Assert.Equal(1, paged.AsPages().Count());
+            var page1 = paged.AsPages().First();
             Assert.Null(page1.Items);
             Assert.Null(page1.NextPageToken);
             // Unnatural things
-            Assert.Empty(paged.Flatten());
-            Assert.Empty(paged.WithFixedPageSize(1));
+            Assert.Empty(paged);
+            Assert.Empty(paged.AsPages().WithFixedSize(1));
         }
 
         [Fact]
@@ -246,13 +264,13 @@ namespace Google.Api.Gax.Rest
             var request = new PageStreamingRequest(server) { PageSize = 0 };
             var paged = server.PagedAsync(request);
             // Natural pages
-            Assert.Equal(1, await paged.Count());
-            var page1 = await paged.First();
+            Assert.Equal(1, await paged.AsPages().Count());
+            var page1 = await paged.AsPages().First();
             Assert.Null(page1.Items);
             Assert.Null(page1.NextPageToken);
             // Unnatural things
-            Assert.Empty(await paged.Flatten().ToArray());
-            Assert.Empty(await paged.WithFixedPageSize(1).ToArray());
+            Assert.Empty(await paged.ToArray());
+            Assert.Empty(await paged.AsPages().WithFixedSize(1).ToArray());
         }
 
         [Fact]
@@ -261,7 +279,7 @@ namespace Google.Api.Gax.Rest
             var server = new FakeServer(s_resourceA, 1);
             var request = new PageStreamingRequest(server) { PageSize = 0 };
             var paged = server.PagedSync(request);
-            var fixedSize = paged.WithFixedPageSize(1);
+            var fixedSize = paged.AsPages().WithFixedSize(1);
             Assert.Throws<NotSupportedException>(() => fixedSize.First());
         }
 
@@ -271,7 +289,7 @@ namespace Google.Api.Gax.Rest
             var server = new FakeServer(s_resourceA, 1);
             var request = new PageStreamingRequest(server) { PageSize = 0 };
             var paged = server.PagedAsync(request);
-            var fixedSize = paged.WithFixedPageSize(1);
+            var fixedSize = paged.AsPages().WithFixedSize(1);
             var ex = await Record.ExceptionAsync(() => fixedSize.First());
             Assert.IsType<NotSupportedException>(ex.InnerException);
         }
